Refactor life loop phases into dedicated modules
